[FIX] payment: Delay payment method line migration
Since other modules create account.payment.method in this migration we
need to wait for them before finding all payment methods and and
connecting them to the journal and acquirer.
